Move anchor link component to shared file
